\name{iMDH-package}
\alias{iMDH-package}
\docType{package}
\title{
  Fast clustering with minimum density hyperplanes.
}
\description{
  Divisive hierarchical clustering based on incrementally estimated minimum density hyperplanes.
}
\details{
  The main functions are imdh and imdh_assign. The imdh function can be used to create or update a clustering solution, while imdh_assign can be used to perform clustering assignments from a given model without updating model parameters.
}
\references{
  Hofmeyr, D. (2021) Clustering Large Data Sets with Incremental Estimation ofLow-density Separating Hyperplanes. \emph{ArXiv preprint}
}
\author{
David P. Hofmeyr
}
